+++ odd-platform-ui/src/components/Search/Filters/FilterItem/MultipleFilterItem/MultipleFilterItemAutocomplete/MultipleFilterItemAutocompleteStyles.ts
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const FilterCount = styled('span')(({ theme }) => ({
+  color: theme.palette.texts.hint,
+}));
